requirements.txt
allure-pytest==2.13.5
Appium-Python-Client==4.0.0
great-expectations==0.18.22
hypothesis==6.142.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
pandas==2.2.3
Pillow==10.4.0
pyautogui==0.9.54
PyInstaller==6.11.0
pynput==1.7.7
pytest==8.4.2
pytest-xdist==3.6.1
pywin32==306
pywinauto==0.6.8
scikit-image==0.24.0
sumy==0.11.0
ttkbootstrap==1.10.1
